+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This CHANGELOG is automatically compiled.
 
 #### ValidationKit
 
+- add support for $anchor in any sub-schema ([f6a5bba](https://github.com/stuartherbert/stusdevkit/commit/f6a5bba)) — Stuart Herbert <stuart@stuartherbert.com>
 - schema $id is now exported to JSON schema ([ba71e0a](https://github.com/stuartherbert/stusdevkit/commit/ba71e0a)) — Stuart Herbert <stuart@stuartherbert.com>
 - add further JSON schema constraints ([3f35f37](https://github.com/stuartherbert/stusdevkit/commit/3f35f37)) — Stuart Herbert <stuart@stuartherbert.com>
 - add support for resolving references during JSON schema import ([ce6db71](https://github.com/stuartherbert/stusdevkit/commit/ce6db71)) — Stuart Herbert <stuart@stuartherbert.com>
@@ -73,7 +74,7 @@ This CHANGELOG is automatically compiled.
 
 #### ValidationKit
 
-- improve round-trip JSON Schema import and export ([3b69aec](https://github.com/stuartherbert/stusdevkit/commit/3b69aec)) — Stuart Herbert <stuart@stuartherbert.com>
+- improve round-trip JSON Schema import and export ([c14fa32](https://github.com/stuartherbert/stusdevkit/commit/c14fa32)) — Stuart Herbert <stuart@stuartherbert.com>
 - $ref handling now matches JSON Schema Draft 2020-12 behaviour ([d45f497](https://github.com/stuartherbert/stusdevkit/commit/d45f497)) — Stuart Herbert <stuart@stuartherbert.com>
 - PHP associative array schemas can now be exported as JSON schema ([7c0bd28](https://github.com/stuartherbert/stusdevkit/commit/7c0bd28)) — Stuart Herbert <stuart@stuartherbert.com>
 - add support for JsonSchema type being an array of types (needed for OpenAPI v3.1 support) ([b599573](https://github.com/stuartherbert/stusdevkit/commit/b599573)) — Stuart Herbert <stuart@stuartherbert.com>